pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "largeforgeai"
version = "0.1.0"
description = "Low-cost LLM training and deployment stack"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "datasets>=2.14.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "trl>=0.7.0",
    "bitsandbytes>=0.41.0",
    "scipy",
    "sentencepiece",
    "protobuf",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
inference = [
    "vllm>=0.2.0",
    "autoawq>=0.1.0",
]

[project.scripts]
llm-train = "llm.training.cli:main"
llm-router = "llm.router.app:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
